Add ability to run Bridge tests from PBN files (#74)
Supports adding *.pbn files under a "Bridge/SAYC" folder in the TestBots project that will automatically get picked up and run as part of the test suite. Only three PBN tags are used for generating tests: Event, Deal, and Auction. The event name is used as the base name of a test and a unique test case is generated for every bid in the action. A full test name therefore looks like "Test Name (Seat: N, Bid: 2)" which would represent a test for North's second bid in the auction.

Co-authored-by: Ted Johnson <johnson520@gmail.com>