Always prioritize bidding Stayman when valid
Introduces a new `Priority` field on `InterpretedBid` which can be used to allow individual bids to indicate their relative priority beyond the default ordering applied to multiple valid bids by `SuggestBid`. This is set to `0` by default, then set to `100` when a bid is interpreted as `Stayman`. Fixes cases where `2NT` was being bid instead of `2C`.