Bridge defensive fixes (take 2) (#90)
Fix #87

In general, the bot should avoid playing high cards immediately on defense to try to force the declaring team to play their high cards. However, there are a number of exceptions where the bot should play high or ruff to avoid losing too many tricks.

This change adds the following additional exceptions to the general rule, so the bot plays high and ruffs more often on defense:
* [x] Ruff in 2nd seat if dummy plays next and has high
* [x] Ruff in 2nd seat if partner is known void in both led suit and trump
* [x] Take tricks quickly in any seat if dummy has long, good side suit
* [x] Take the last sure tricks to defeat the contract from any seat

This version fixes a "Sequence contains no element" error that happened when dummy was void in the led suit. It also updates the fuzzer to ensure we catch similar conditions in the future (the fuzzer previously wasn't removing cards from hands so it played duplicate cards and didn't cover players becoming void).