Fix more 2nd and 3rd seat defensive plays
Notably this fixes cases where 3rd seat would play high, but not try to win the trick. It now more correctly plays as high as necessary, meaning it will try to win the trick but will watch for cases where it is safe to play a lower card and still win the trick.

Play from 2nd seat has also been improved to pick fall-through when trying to win tricks quickly if we're not holding any high cards. It also handles a new special-case where we should play an Ace if dummy follows and will be forced to play a singleton King.

Lastly, logic for trumping in has been simplified to do so in most cases unless clearly not beneficial.