Bridge: More defensive play improvements (#95)
Fix #91
Ref #94

Notably this fixes cases where 3rd seat would play high, but not try to win the trick. The bot now more correctly plays as high as necessary, meaning it will try to win the trick but will watch for cases where it is safe to play a lower card and still win the trick.

Play from 2nd seat has also been improved to fall-through when trying to win tricks quickly if not holding any high cards. It also handles a new special-case of playing an Ace if dummy follows and will be forced to play a singleton King.

Lastly, logic for ruffing on defense has been simplified to do so in most cases when possible unless clearly not beneficial.

Also fixes some test infrastructure bugs where the state being provided wasn't quite correct:
* Match sort of cards in tests with production
* Fix assigning bids to correct seats in tests
* Don't overwrite player seats in tests if already set
* Mod declarer seat by 4 when generating fuzz tests
* Fix calculating trick taker in TestCardState