Don't count off-suit boss cards if trick contains trump
This happened because the bot knew LHO was void, but failed to take into account trump had already been played to the trick. It also failed to take into account the need to follow suit for the current trick.

Fixed by excluding off-suits from sure tricks if the current trick contains trump AND only moving forward with sure tricks if at least one of the winners is in the led suit.