Bridge: Fix calculating sure tricks when current trick has trump (#130)
Fix #129

This happened because the bot knew LHO was void but failed to take into account trump had already been played to the trick. It also failed to take into account the led suit for the current trick (as a card would only win if it followed suit).

Fixed by excluding off-suits from sure tricks if the current trick contains trump AND only moving forward with sure tricks if at least one of the winners is in the led suit.

* Fix an error calculating declarer's cards in trump

* Fix having tests show declarer's hand to dummy

Needed when we're testing play from dummy as dummy is actually played by the declarer.