Bridge: Fix bids missing HandShape requirements
Found via a challenge this past weekend where the bot made a bid indicating it had 5+ Spades when it only actually had 3.

When investigating, I found this occurred because even though "5+ Spades" was listed in the Description, no HandShape was actually set. Fixed by setting the HandShape.

I decided to search the code for similar bugs in other places and found a few more (which I also fixed). These resulted in a couple of changes to the results in the SAYC test suite which gave us some better bids (though still not the ideal ones in those two cases).