Add failing test where bot didn't duck
I've debugged this into GoodCardToGiveToPartner.

The bot is trying to "give" the QS to their partner because it's worth points toward game and isn't known to be the current highest trump card (so in theory it could be captured).

However this is actually wasting a high-enough card that could take a future trick, especially since the bot has enough cards to play low if the KS came out on a future trick.

Need to think a bit about what type of exception to make here (possibly to always exclude honors in trump from this logic...).