Add more failing tests
When choosing to trump in or play over an existing trump, the current logic plays the highest trump if it will win, or lowest trump otherwise.

This needs nuance added based on the seat the player is playing from. Should generally play low from 2nd seat if we don't have boss and high from 3rd regardless. And from 4th we should only play as high as necessary and prefer playing a pointer.