Pitch: Fix some bad plays (#148)
Fix #51
Fix #127
Fix #132 

* Add failing test where bot didn't duck

I've debugged this into GoodCardToGiveToPartner.

The bot is trying to "give" the QS to their partner because it's worth points toward game and isn't known to be the current highest trump card (so in theory it could be captured).

However this is actually wasting a high-enough card that could take a future trick, especially since the bot has enough cards to play low if the KS came out on a future trick.

Need to think a bit about what type of exception to make here (possibly to always exclude honors in trump from this logic...).

* Add more failing tests

When choosing to trump in or play over an existing trump, the current logic plays the highest trump if it will win, or lowest trump otherwise.

This needs nuance added based on the seat the player is playing from. Should generally play low from 2nd seat if we don't have boss and high from 3rd regardless. And from 4th we should only play as high as necessary and prefer playing a pointer.

* More tests

* Update bot to pass new tests

* Handle potential null

* Fix more cases from last seat

* More tests

* Fix leading high trump when on pitching team

* Fix priority for dumping cards

* Minor bidding refinement

* Remove redundant check

* Fix preferring discarding non-trump

* Simplify check