Bridge: Stop passing when too strong
Fix #167

Previous/y if no bids matched because the hand was too strong for any of them (and didn't fit the shape of others) the bot would pass unless partner's bid was forcing.

This change uncaps bids when this situation is hit allowing the bot to make a bid it was otherwise too strong for (which is at least better than passing, if not always optimal).