Update BaseBot.CanSeeHand to derive from cards
Updates `CanSeeHand` to return `true` if the target both has cards and all those cards are not of unknown suit or rank.

The expectation is that a caller won't pass cards as part of the suggest method that the bot isn't supposed to be able to see.

Note that this returns `false` if the target has no cards, mostly to accommodate existing tests that don't pass in a set of unknown cards for face-down hands.

Typically callers should pass the correct number of unknown cards for a face-down hand as bots are allowed to count the number of cards remaining in a player's hand.