500: Bid only as high as needed to win the game
Fix #154

This take into account whether the bot's team is currently winning the bid, whether the opponents may still overbid, and whether we'll be able to bid again if the opponents do overbid.

The bot will stop bidding once the value of the bid is sufficiently high to win the game, but will keep raising the opponents if they compete up until the original bid the bot would have made.