Improve playing alone bids with gaps
Updates the bot to keep playing high trump and off-suit cards even if they're not the highest potentially in play so long as the opponents haven't taken a trick yet.

This helps ensure the bot doesn't miss opportunities to take all 5 tricks when the high Jack or an off-suit Ace are out of play.