Bridge: Minimize leading trump on defense
While there are good times to do this, it isn't common. The bot plays more realistically if we limit this to more obvious cases.

Now the bot will only lead trump on defense if partner is void in trump, both declarer and dummy still have trump, and the bot is holding the highest remaining trump.